validation-api/pyproject.toml
[tool.poetry]
name = "validation-api"
version = "0.1.0"
description = ""
authors = ["amondal2 <ab59@cdc.gov>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
attrs = "24.2.0"
blinker = "1.8.2"
certifi = "2024.8.30"
cfgv = "3.4.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
colorama = "0.4.6"
detect-secrets = "1.5.0"
distlib = "0.3.8"
filelock = "3.15.4"
flask = "3.0.3"
identify = "2.6.0"
idna = "3.8"
itsdangerous = "2.2.0"
jinja2 = "3.1.4"
jsonschema = "4.23.0"
jsonschema-specifications = "2023.12.1"
markupsafe = "2.1.5"
nodeenv = "1.9.1"
platformdirs = "4.2.2"
pre-commit = "3.8.0"
pyyaml = "6.0.2"
referencing = "0.35.1"
requests = "2.32.3"
rpds-py = "0.20.0"
urllib3 = "2.2.2"
virtualenv = "20.26.3"
werkzeug = "3.0.4"
pytest = "^8.3.3"
azure-identity = "^1.17.1"
azure-keyvault = "^4.2.0"
toml = "^0.10.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "validation-api"
